desc/requirements.txt
anndata==0.10.7
scanpy==1.10.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
umap-learn==0.5.6
matplotlib==3.8.4
seaborn==0.13.2
torch==2.3.1